requirements.txt
fastapi
uvicorn[standard]
prisma
psycopg2-binary
python-docx
PyPDF2
google-generativeai
python-dotenv
reportlab
python-docx
